+++ src/core/request/engines/fetch.ts
@@ -65,7 +65,7 @@ const request: RequestEngine = (params) => {
 			{status} = await AbortablePromise.resolve(isOnline(), p.parent);
 
 		if (!status) {
-			return reject(new RequestError('offline', {
+			return reject(new RequestError(RequestError.Offline, {
 				request: <NormalizedCreateRequestOptions>normalizedOpts
 			}));
 		}
@@ -122,7 +122,7 @@ const request: RequestEngine = (params) => {
 		}, (error) => {
 			clearTimeout(timer);
 
-			const type = error.name === 'AbortError' ? 'timeout' : 'engine';
+			const type = error.name === 'AbortError' ? RequestError.Timeout : RequestError.Engine;
 			reject(new RequestError(type, {error}));
 		});